Add support for phases to sriov-config-daemon service
The service has required argument "phase"
Two phases are supported now:
 * pre - before the NetworkManager or systemd-networkd
* post - after the NetworkManager or systemd-networkd
"sriov-config" systemd unit is responsible for starting the service in the "pre" phase mode.
"sriov-config-post-network" systemd unit starts the service in the "post" phase mode.
The service may use different plugins for each phase and call different initialization flows.
The "post" phase checks the completion status of the "pre" phase by reading the sriov result file.
The "pre" phase should set "InProgress" status if it succeeds or "Failed" otherwise.
If the result of the "pre" phase is different than "InProgress", then the "post" phase will not be executed
and the execution result will be forcefully set to "Failed".

Signed-off-by: Yury Kulazhenkov <ykulazhenkov@nvidia.com>